requirements.txt
# Core Framework
streamlit==1.31.0
langchain==0.1.9
langchain-core==0.1.23
langchain-google-genai==0.0.9

# AI/ML Libraries
google-generativeai==0.3.2
langchain-community==0.0.24

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Database
sqlalchemy==2.0.25

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.2
altair==5.2.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0

# Data Validation
pandera==0.18.0

# Additional Utilities
openpyxl==3.1.2
pyarrow==15.0.0
